split FileLibraryManager into generic LibraryManager and FileLibraryStorage
